Endpoint Update to accept XML Request
This endpoint searches for transactions in the PostgreSQL database using a keyword provided as a URL parameter. It calls the stored procedure 'search_table2' with the keyword and stores the results in 'response_data'. The response format is determined by the 'Accept' header: XML if it's 'application/xml', otherwise JSON, both with a 200 status code. Exceptions during the query result in a 500 status code with an error message formatted according to the 'Accept' header.